added register component
